requirements.txt
django==2.2.3
djangorestframework==3.10.1
pylint==2.3.1
PyJWT==1.7.1
psycopg2-binary==2.8.2
python-dotenv==0.10.3
cloudinary==1.16.0
